RelativePosition added - addToLineAnnotation x y params changed to doubles (for single decimal precision)

+++ src/nl/ou/jp/domain/SlideShowService.java
@@ -30,7 +30,7 @@ public interface SlideShowService {
 	
 	void startLineAnnotation(int index, int lineWeight, int color);
 	
-	void addToLineAnnotation(int index, int x, int y);
+	void addToLineAnnotation(int index, double x, double y);
 	
 	SlideShowEventDispatcher getSlideShowEventDispatcher();
 	SlideEventDispatcher getSlideEventDispatcher();
